chore: simplify link modification code
